Fix uninitialized data and unused functions

+++ tests/document.c
@@ -108,7 +108,7 @@ START_TEST(test_document_set_number_of_pages) {
   fail_unless(zathura_document_set_number_of_pages(NULL, 1) == ZATHURA_ERROR_INVALID_ARGUMENTS);
 
   /* valid arguments */
-  size_t number_of_pages;
+  unsigned int number_of_pages;
   fail_unless(zathura_document_get_number_of_pages(document, &number_of_pages) == ZATHURA_ERROR_OK);
   fail_unless(zathura_document_set_number_of_pages(document, number_of_pages) == ZATHURA_ERROR_OK);
 } END_TEST
